+++ examples/resources/prodvana_k8s_runtime/resource.tf
@@ -1,3 +1,13 @@
 resource "prodvana_k8s_runtime" "example" {
   name = "my-k8s-runtime"
+  labels = [
+    {
+      label = "env"
+      value = "staging"
+    },
+    {
+      label = "region"
+      value = "us-central1"
+    },
+  ]
 }